feat: add better system env support
This PR adds support for a --load-system-env flag. This works similar to the existing
--load-local-env variable on the start command but is required to be used in tandem with the --env
flag and will use the file specified in --env as a filter. Only empty values in the env file will be
populated by the system env variables. This PR doesn't add that behavior to the "start" command as
this would be a breaking change that will be introduced in a separate version.

fix #144